Make all libraries static
This allows me to use the --static-swift-stdlib flag when
building my project, which results in a single binary containing
every Swift package as well as the Foundation and Swift runtime.
Having multiple dynamic libraries would make this impossible, but
the only thing keeping them dynamic is because that is how the
upstream maintainer did it when I forked.